created ReplyTweet modal, form and button with logic
