+++ src/components/ReplyTweetButton/index.tsx
@@ -1,0 +1,7 @@
+import ModalTweetReplyForm from "../../containers/ModalTweetReplyForm";
+
+const ReplyTweet: React.FC = () => {
+  return <ModalTweetReplyForm />;
+};
+
+export default ReplyTweet;
