+++ src/components/TweetReplyForm.tsx
@@ -1,0 +1,81 @@
+import { EmojiSmile, Image as ImageIcon } from "react-bootstrap-icons";
+import { useTranslation } from "react-i18next";
+import styled from "styled-components";
+
+const StyledIcon = styled.div`
+  color: rgb(0, 0, 255, 0.6);
+  &:hover {
+    background-color: rgb(0, 0, 255, 0.1);
+    cursor: pointer;
+  }
+`;
+const StyledInput = styled.input`
+  width: 100%;
+  margin-top: 10px;
+  border: none;
+`;
+
+const StyledDiv = styled.div`
+  width: 100%;
+`;
+
+const TweetReplyForm: React.FC = () => {
+  const { t } = useTranslation();
+
+  const handleTweetButton = (): void => {
+    console.warn("Tweeted");
+  };
+  const handleImgUpload = (): void => {
+    console.warn("Img Upload");
+  };
+  const handleEmojiPaste = (): void => {
+    console.warn("Emoji paste");
+  };
+  return (
+    <div>
+      <div className="border-0 p-3 d-flex text-start justify-content-start">
+        <div className="">
+          <img
+            className="rounded-circle"
+            src="http://dummyimage.com/105x100.png/5fa2dd/ffffff"
+            width="48px"
+            height="48px"
+            alt="avatar"
+          />
+        </div>
+        <div className="flex-grow-1 ms-2">
+          <StyledDiv>Xiaomi</StyledDiv>
+          <StyledDiv>@Xiaomi</StyledDiv>
+          <br />
+          <StyledDiv>
+            "Phasellus sit amet erat. undefineda tempus. Vivamus in felis eu
+            sapien cursus vestibulum.\n\nProin eu mi. undefineda ac enim. In
+            tempor, turpis nec euismod scelerisque, quam turpis adipiscing
+            lorem, vitae mattis nibh ligula nec sem.\n\nDuis aliquam convallis
+            nunc. Proin at turpis a pede posuere nonummy. Integer non velit."
+          </StyledDiv>
+          <br />
+          <StyledInput placeholder={t("tweet your reply")} />
+          <div className="d-flex justify-content-between">
+            <div className="d-flex p-2">
+              <StyledIcon className="rounded-circle" onClick={handleImgUpload}>
+                <ImageIcon className="m-2 fs-5" />
+              </StyledIcon>
+              <StyledIcon className="rounded-circle" onClick={handleEmojiPaste}>
+                <EmojiSmile className="m-2 fs-5" />
+              </StyledIcon>
+            </div>
+            <button
+              className="btn btn-primary rounded-5 d-flex align-items-center m-2"
+              onClick={handleTweetButton}
+            >
+              {t("translation:Reply")}
+            </button>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default TweetReplyForm;
